==> code/05-Generating-Coverage-Tracks.ipynb <==
# Generating Coverage Tracks

In order to visualize my DML tracks in IGV, I need to match these features to the actual sample tracks. Since they are only 1x coverage, bedGraphs will not work. I will generate 5x tracks for all sample coverage files so I can use them in IGV.

Methods:

0. Prepare for Analyses
1. Obtain Coverage Files
3. Create 5x Bedgraphs

## 0. Prepare for Analyses

### 0a. Set Working Directory

pwd

cd ../analyses/

!mkdir 2019-03-07-IGV-Verification

cd 2019-03-07-IGV-Verification/

## 1. Obtain Coverage Files

The file are in [this folder](http://gannet.fish.washington.edu/spartina/2018-10-10-project-virginica-oa-Large-Files/2018-11-07-Bismark-Mox/). I'll use `wget` to download them.

#Download files from gannet. The files will be downloaded in the same directory structure they are in online.
!wget -r -l1 --no-parent -A.deduplicated.bismark.cov.gz \
http://gannet.fish.washington.edu/spartina/2018-10-10-project-virginica-oa-Large-Files/2018-11-07-Bismark-Mox/

#Move all files from gannet folder to the current directory
!mv gannet.fish.washington.edu/spartina/2018-10-10-project-virginica-oa-Large-Files/2018-11-07-Bismark-Mox/* .

#Confirm all files were moved
!ls

#Remove the empty gannet directory
!rm -r gannet.fish.washington.edu

#Unzip the coverage files
!gunzip *cov.gz

#Confirm files were unzipped
!ls *cov

#See what the file looks like
!head -n 1 zr2096_10_s1_R1_val_1_bismark_bt2_pe.deduplicated.bismark.cov

## Create 5x Tracks

I will replicate the above process to get tracks with 5x coverage. Claire and Mac have used 5x coverage, so I want to see what my data looks like here.

### Percent Methylation Only

%%bash
for f in *.cov
do
    awk '{print $1, $2-1, $2, $4, $5+$6}' ${f} | awk '{if ($5 >= 5) { print $1, $2-1, $2, $4 }}' \
> ${f}_5x.bedgraph
done

#Confirm 5x tracks were created
!ls *5x.bedgraph

!head zr2096_7_s1_R1_val_1_bismark_bt2_pe.deduplicated.bismark.cov_5x.bedgraph

### Coverage and Percent Methylation

%%bash
for f in *.cov
do
    awk '{print $1, $2-1, $2, $4, $5+$6}' ${f} | awk '{if ($5 >= 5) { print $1, $2-1, $2, $4, $5 }}' \
> ${f}_5x.percentMeth.cov
done

#Confirm 5x tracks were created
!ls *5x.percentMeth.cov

!head zr2096_7_s1_R1_val_1_bismark_bt2_pe.deduplicated.bismark.cov_5x.percentMeth.cov
