==> code/07-Generating-Genome-Feature-Tracks.ipynb <==
# Generating Genome Feature Tracks

I will create genome feature tracks to use in downstream analyses. While pre-made genome feature tracks exist, it's beneficial to generate these tracks to understand what elements they contain. My code was modified from the code provided in the [FROGER Github repository](https://github.com/hputnam/FROGER/wiki/Genome-Sequence-Files-and-Feature-Tracks).

1. Download *C. virginica* genome file
2. Separate various tracks
3. Visualize tracks in IGV
4. Characterize track overlap with CG motifs

## 0. Set working directory

pwd

cd ../genome-feature-tracks/

## 1. Download files

### 1a. Pre-generated CG motif track

!curl http://eagle.fish.washington.edu/Cvirg_tracks/C_virginica-3.0_CG-motif.bed > C_virginica-3.0_CG-motif.bed

!head C_virginica-3.0_CG-motif.bed

### 1b. *C. virgincia* genome from NCBI

!curl ftp://ftp.ncbi.nlm.nih.gov/genomes/Crassostrea_virginica/GFF/ref_C_virginica-3.0_top_level.gff3.gz > ref_C_virginica-3.0_top_level.gff3.gz

!gunzip ref_C_virginica-3.0_top_level.gff3.gz

!head ref_C_virginica-3.0_top_level.gff3

## 2. Set variable paths

bedtoolsDirectory = "/Users/Shared/bioinformatics/bedtools2/bin/"

fullGenome = "../data/C_virginica-3.0_genomic.fa"

CGMotifList = "C_virginica-3.0_CG-motif.bed"

chromLengths = "2018-06-15-bedtools-Chromosome-Lengths.txt"

!cut -f1 {chromLengths} > 2019-05-28-bedtools-Chromosome-Names.txt

!cat 2019-05-28-bedtools-Chromosome-Names.txt

chromNames = "2019-05-28-bedtools-Chromosome-Names.txt"

## 3. Generate feature tracks

### 3a. Genes

#Isolate gene entries. Tab must be included between "Gnomon" and "gene"
!grep "Gnomon	gene" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_gene_yrv.gff3

#Sort file for downstream use
!{bedtoolsDirectory}sortBed \
-faidx {chromNames} \
-i C_virginica-3.0_Gnomon_gene_yrv.gff3 \
> C_virginica-3.0_Gnomon_gene_sorted_yrv.gff3

#Set variable path
geneList = "C_virginica-3.0_Gnomon_gene_sorted_yrv.gff3"

#View file
!head {geneList}

#Count the number of genes
!wc -l {geneList}
!echo "genes in the C. virginica genome"

%%bash
#GFF files can be a lot to visualize, so I'll create a BEDfile version too
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_gene_sorted_yrv.gff3 \
> C_virginica-3.0_Gnomon_gene_sorted_yrv.bed

### 3b. mRNA

!grep "Gnomon	mRNA" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_mRNA_yrv.gff3

mRNAList = "C_virginica-3.0_Gnomon_mRNA_yrv.gff3"

!head -n1 {mRNAList}

!wc -l {mRNAList}
!echo "mRNAs in the C. virginica genome"

%%bash
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_mRNA_yrv.gff3 \
> C_virginica-3.0_Gnomon_mRNA_yrv.bed

### 3c. Exons

!grep "Gnomon	exon" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_exon_yrv.gff3

#Sort the exon file for downstream use
!{bedtoolsDirectory}sortBed \
-faidx {chromNames} \
-i C_virginica-3.0_Gnomon_exon_yrv.gff3 \
> C_virginica-3.0_Gnomon_exon_sorted_yrv.gff3

exonList = "C_virginica-3.0_Gnomon_exon_sorted_yrv.gff3"

!head {exonList}

!wc -l {exonList}
!echo "exons in the C. virginica genome"

%%bash
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_exon_sorted_yrv.gff3 \
> C_virginica-3.0_Gnomon_exon_sorted_yrv.bed

### 3d. Intergenic regions

By definition, these are regions that aren't genes. I can use `complementBed` to find all regions that aren't genes, and `subtractBed` to remove exons and create this track.

!{bedtoolsDirectory}complementBed \
-i {geneList} -sorted \
-g {chromLengths} \
| {bedtoolsDirectory}subtractBed \
-a - \
-b {exonList} \
> C_virginica-3.0_Gnomon_intergenic_yrv.gff3

intergenicList = "C_virginica-3.0_Gnomon_intergenic_yrv.gff3"

!head {intergenicList}

!wc -l {intergenicList}
!echo "intergenic regions in the C. virginica genome"

%%bash
awk '{print $1"\t"$2"\t"$3}' C_virginica-3.0_Gnomon_intergenic_yrv.gff3 \
> C_virginica-3.0_Gnomon_intergenic_yrv.bed

### 3e. Coding sequences

!grep "Gnomon	CDS" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_CDS_yrv.gff3

#Sort the exon file for downstream use
!{bedtoolsDirectory}sortBed \
-faidx {chromNames} \
-i C_virginica-3.0_Gnomon_CDS_yrv.gff3 \
> C_virginica-3.0_Gnomon_CDS_sorted_yrv.gff3

CDSList = "C_virginica-3.0_Gnomon_CDS_sorted_yrv.gff3"

!head {CDSList}

!wc -l {CDSList}
!echo "CDSRNA in the C. virginica genome"

%%bash
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_CDS_sorted_yrv.gff3 \
> C_virginica-3.0_Gnomon_CSD_sorted_yrv.bed

### 3f. Non-coding Sequences

I can use `complementBed` to create a non-coding sequence track. This track can then be used to create an intron track.

!{bedtoolsDirectory}complementBed \
-i {exonList} \
-g 2018-06-15-bedtools-Chromosome-Lengths.txt \
> C_virginica-3.0_Gnomon_noncoding_yrv.gff3

nonCDS = "C_virginica-3.0_Gnomon_noncoding_yrv.gff3"

!head {nonCDS}

!wc -l {nonCDS}
!echo "non-coding sequences in the C. virginica genome"

%%bash
awk '{print $1"\t"$2"\t"$3}' C_virginica-3.0_Gnomon_noncoding_yrv.gff3 \
> C_virginica-3.0_Gnomon_noncoding_yrv.bed

### 3g. Introns

#The intersections betwen the non-coding sequences and genes are by definition introns
!{bedtoolsDirectory}intersectBed \
-a {nonCDS} \
-b {geneList} -sorted \
> C_virginica-3.0_Gnomon_intron_yrv.gff3

intronList = "C_virginica-3.0_Gnomon_intron_yrv.gff3"

!head {intronList}

!wc -l {intronList}
!echo "introns in the C. virginica genome"

%%bash
awk '{print $1"\t"$2"\t"$3}' C_virginica-3.0_Gnomon_intron_yrv.gff3 \
> C_virginica-3.0_Gnomon_intron_yrv.bed

### 3h. lncRNA

Note: there are only lncRNA and not the associated exons.

!grep "Gnomon	lnc_RNA" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_lncRNA_yrv.gff3

lncRNAList = "C_virginica-3.0_Gnomon_lncRNA_yrv.gff3"

!head {lncRNAList}

!wc -l {lncRNAList}
!echo "lncRNA in the C. virginica genome"

%%bash
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_lncRNA_yrv.gff3 \
> C_virginica-3.0_Gnomon_lncRNA_yrv.bed

### 3i. Untranslated regions of exons

These can be derived by subtracting coding sequences and exons.

!{bedtoolsDirectory}subtractBed \
-a {exonList} \
-b {CDSList} \
-sorted \
-g {chromLengths} \
> C_virginica-3.0_Gnomon_exonUTR_yrv.gff3

exonUTR = "C_virginica-3.0_Gnomon_exonUTR_yrv.gff3"

!head {exonUTR}

!wc -l {exonUTR}
!echo "untranslated regions of exons in the C. virginica genome"

%%bash
awk '{print $1"\t"$4"\t"$5}' C_virginica-3.0_Gnomon_exonUTR_yrv.gff3 \
> C_virginica-3.0_Gnomon_exonUTR_yrv.bed

### 3j. mtDNA

The smallest chromosome is mitochondrial DNA. I can create a separate track for mtDNA by simply using `grep`.

!grep "NC_007175.2" ref_C_virginica-3.0_top_level.gff3 > C_virginica-3.0_Gnomon_mtDNA_yrv.gff3

mtDNA = "C_virginica-3.0_Gnomon_mtDNA_yrv.gff3"

!head {mtDNA}

!wc -l {mtDNA}
!echo "mitchondrial DNA sequences in the *C. virginica* genome"

## 4. Visualize in IGV

The best way to confirm that I created my tracks correctly is to look at them in the Integrated Genome Viewer (IGV). Visualization can be done at [this link](https://github.com/fish546-2018/yaamini-virginica/blob/master/analyses/2019-05-13-Generating-Genome-Feature-Tracks/2019-05-13-Genome-Track-Verification.xml).

## 5. Characterize CG motif locations

#Roughly count the number of Cs in the full genome
!fgrep -o -i C {fullGenome} | wc -l

#Count the number of CG motifs in the premade file
!wc -l {CGMotifList}

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {geneList} \
| wc -l
!echo "CG motifs overlap with genes"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {mRNAList} \
| wc -l
!echo "CG motifs overlap with mRNA"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {exonList} \
| wc -l
!echo "CG motifs overlap with exons"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {intergenicList} \
| wc -l
!echo "CG motifs overlap with intergenic regions"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {CDSList} \
| wc -l
!echo "CG motifs overlap with coding sequences"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {nonCDS} \
| wc -l
!echo "CG motifs overlap with non-coding sequences"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {intronList} \
| wc -l
!echo "CG motifs overlap with introns"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {lncRNAList} \
| wc -l
!echo "CG motifs overlap with lncRNA"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {exonUTR} \
| wc -l
!echo "CG motifs overlap with untranslated regions of exons"

!{bedtoolsDirectory}intersectBed \
-u \
-a {CGMotifList} \
-b {mtDNA} \
| wc -l
!echo "CG motifs overlap with mitochondrial DNA"
