==> code/09-Characterizing-CpG-Methylation.ipynb <==
# Characterizing CpG Methylation

To describe general metylation trends, irrespective of pCO<sub>2</sub> treatment in *C. virginica* gonad sequence data, I need to characterize individual CpG loci. Gavery and Roberts (2013) and Olson and Roberts (2013) define a CpG locus as methylated if at least half of the reads remained unconverted after bisulfite treatment. I will use information in a master `.cov` file to identify methylated CpG loci.

Another thing I will do is identify methylation islands by replicating [Jeong et al. 2018](https://academic.oup.com/gbe/article/10/10/2766/5098531). I will use [their script](https://github.com/soojinyilab/Methylation-Islands) but modify parameters to reflect differences in insect and *C. virginica* methylation.

1. Download coverage file
2. Limit to 5x coverage
3. Characterize methylation levels for loci
4. Characterize loci locations
5. Identify methylation islands

## 0. Prepare for analyses

## 0a. Set working directory

pwd

cd ../analyses/

!mkdir 2019-03-18-Characterizing-CpG-Methylation

cd 2019-03-18-Characterizing-CpG-Methylation/

## 1. Obtain coverage files

#Download file from gannet. This file is a concatenation of coverage and methylation information for all samples
!wget http://gannet.fish.washington.edu/Atumefaciens/20190312_cvir_gonad_bismark/total_reads_bismark/cvir_bsseq_all_pe_R1_bismark_bt2_pe.bismark.cov.gz

#Unzip the coverage file
!gunzip *cov.gz

#Confirm file was unzipped
!ls *cov

#See what the file looks like. 
#Columns: <chromosome> <start position> <end position> <methylation percentage> <count methylated> <count unmethylated>
!head cvir_bsseq_all_pe_R1_bismark_bt2_pe.bismark.cov

#See how many loci have data
!awk '{if ($5+$6 >= 1) { print $1, $2-1, $3, $4, $5+$6}}' cvir_bsseq_all_pe_R1_bismark_bt2_pe.bismark.cov \
| wc -l

14,026,131 CpGs have data, which is close to the 14,458,703 CG motifs in the *C. virginica* genome.

## 2. Limit to 5x coverage

#If total coverage (count methylated + unmethylated) is greater than 5
#then print the chromosome, start pos -1, stop pos, percent methylation, and total coverage
#Save output as new file
!awk '{if ($5+$6 >= 5) { print $1, $2-1, $3, $4, $5+$6}}' cvir_bsseq_all_pe_R1_bismark_bt2_pe.bismark.cov \
> 2019-04-09-All-5x-CpGs.bedgraph

#Check columns for one of the file: <chromosome> <start position> <stop position> <percent methylation> <coverage>
!head 2019-04-09-All-5x-CpGs.bedgraph

#Count loci with 5x coverage
!wc -l 2019-04-09-All-5x-CpGs.bedgraph

I have data for 4,304,257 CpG loci with 5x coverge.

#Replace delimiters to save .bedgraph as .csv
!awk '{print $1","$2","$3","$4 }' 2019-04-09-All-5x-CpGs.bedgraph \
> 2019-04-09-All-5x-CpGs.csv

#Confirm .csv creation
!head 2019-04-09-All-5x-CpGs.csv

## 3. Characterize methylation levels for loci

Olson and Roberts (2014) define the following categories for CpG methylation:

- Methylated (50% methylation and above)
- Sparsely methylated (0-50% methylated)
- Unmethylated (0% methylation)

I will slightly modify this since I have multiple samples:

- Methylated (50% methylation and above)
- Sparsely methylated (10-50% methylated)
- Unmethylated (10% methylation and below)

### 3a. Methylated loci

#If percent methylation is greater or equal to 50, then save the loci information
!awk '{if ($4 >= 50) { print $1, $2, $3, $4 }}' 2019-04-09-All-5x-CpGs.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Methylated.bedgraph

#Confirm methylated loci were saved
!head 2019-04-09-All-5x-CpG-Loci-Methylated.bedgraph

#Count methylated loci
!wc -l 2019-04-09-All-5x-CpG-Loci-Methylated.bedgraph

#Replace delimiters to save .bedgraph as .csv
!awk '{print $1","$2","$3","$4 }' 2019-04-09-All-5x-CpG-Loci-Methylated.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Methylated.csv

#Check .csv was saved
!head 2019-04-09-All-5x-CpG-Loci-Methylated.csv

### 3b. Sparsely methylated loci

%%bash
awk '{if ($4 < 50) { print $1, $2, $3, $4}}' 2019-04-09-All-5x-CpGs.bedgraph \
| awk '{if ($4 > 10) { print $1, $2, $3, $4 }}' \
> 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bedgraph

#Confirm sparsely methylated loci were saved
!head 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bedgraph

#Count sparsely methylated loci
!wc -l 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bedgraph

### 3c. Unmethylated loci

!awk '{if ($4 <= 10) { print $1, $2, $3, $4 }}' 2019-04-09-All-5x-CpGs.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Unmethylated.bedgraph

#Confirm unmethylated loci were saved
!head 2019-04-09-All-5x-CpG-Loci-Unmethylated.bedgraph

#Count unmethylated loci
!wc -l 2019-04-09-All-5x-CpG-Loci-Unmethylated.bedgraph

## 4. Characterize loci locations

My final step is to characterize the location of various loci categories in the genome. I will use `intersectBed` to find overlaps between all 5x CpGs, methylated loci, sparsely methylated loci, and unmethylated loci with exons, introns, mRNA coding regions, transposable elements, and putative promoter regions.

### 4a. Create `.bed` files

#### All 5x CpGs

%%bash
awk '{print $1"\t"$2"\t"$3}' 2019-04-09-All-5x-CpGs.bedgraph \
> 2019-04-09-All-5x-CpGs.bed

#Confirm file creation
!head 2019-04-09-All-5x-CpGs.bed

#### Methylated loci

%%bash
awk '{print $1"\t"$2"\t"$3}' 2019-04-09-All-5x-CpG-Loci-Methylated.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Methylated.bed

#Confirm file creation
!head 2019-04-09-All-5x-CpG-Loci-Methylated.bed

#### Sparsely methylated loci

%%bash
awk '{print $1"\t"$2"\t"$3}' 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bed

#Confirm file creation
!head 2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bed

#### Unmethylated loci

%%bash
awk '{print $1"\t"$2"\t"$3}' 2019-04-09-All-5x-CpG-Loci-Unmethylated.bedgraph \
> 2019-04-09-All-5x-CpG-Loci-Unmethylated.bed

#Confirm file creation
!head 2019-04-09-All-5x-CpG-Loci-Unmethylated.bed

### 4b. Set variable paths

bedtoolsDirectory = "/Users/Shared/bioinformatics/bedtools2/bin/"

all5xCpGs = "2019-04-09-All-5x-CpGs.bed"

methylatedLoci = "2019-04-09-All-5x-CpG-Loci-Methylated.bed"

sparselyMethylatedLoci = "2019-04-09-All-5x-CpG-Loci-Sparsely-Methylated.bed"

unmethylatedLoci = "2019-04-09-All-5x-CpG-Loci-Unmethylated.bed"

putativePromoters = "../../genome-feature-tracks/2019-05-29-Flanking-Analysis/2019-05-29-mRNA-Promoter-Track.bed"

exonUTR = "../../genome-feature-tracks/C_virginica-3.0_Gnomon_exonUTR_yrv.gff3"

exonList = "../../genome-feature-tracks/C_virginica-3.0_Gnomon_exon_sorted_yrv.bed"

intronList = "../../genome-feature-tracks/C_virginica-3.0_Gnomon_intron_yrv.bed"

geneList = "../../genome-feature-tracks/C_virginica-3.0_Gnomon_gene_sorted_yrv.bed"

transposableElementsAll = "../../genome-feature-tracks/C_virginica-3.0_TE-all.gff"

intergenic = "../../genome-feature-tracks/C_virginica-3.0_Gnomon_intergenic_yrv.gff3"

### 4c. Exons

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {exonList} \
| wc -l
!echo "all 5x CpG loci overlaps with exons"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {exonList} \
> 2019-05-29-All5xCpGs-Exon.txt

!head 2019-05-29-All5xCpGs-Exon.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {exonList} \
| wc -l
!echo "methylated loci overlaps with exons"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {exonList} \
> 2019-05-29-MethLoci-Exon.txt

!head 2019-05-29-MethLoci-Exon.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {exonList} \
| wc -l
!echo "sparsely methylated loci overlaps with exons"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {exonList} \
> 2019-05-29-SparseMethLoci-Exon.txt

!head 2019-05-29-SparseMethLoci-Exon.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {exonList} \
| wc -l
!echo "unmethylated loci overlaps with exons"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {exonList} \
> 2019-05-29-UnMethLoci-Exon.txt

!head 2019-05-29-UnMethLoci-Exon.txt

### 4d. Introns

#### All 5x CpG

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {intronList} \
| wc -l
!echo "all 5x CpG loci overlaps with introns"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {intronList} \
> 2019-05-29-All5xCpGs-Intron.txt

!head 2019-05-29-All5xCpGs-Intron.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {intronList} \
| wc -l
!echo "methylated loci overlaps with introns"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {intronList} \
> 2019-05-29-MethLoci-Intron.txt

!head 2019-05-29-MethLoci-Intron.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {intronList} \
| wc -l
!echo "sparsely methylated loci overlaps with introns"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {intronList} \
> 2019-05-29-SparseMethLoci-Intron.txt

!head 2019-05-29-SparseMethLoci-Intron.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {intronList} \
| wc -l
!echo "unmethylated loci overlaps with introns"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {intronList} \
> 2019-05-29-UnMethLoci-Intron.txt

!head 2019-05-29-UnMethLoci-Intron.txt

### 4e. Exon UTR

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {exonUTR} \
| wc -l
!echo "all 5x CpG loci overlaps with exon UTR"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {exonUTR} \
> 2019-05-29-All5xCpGs-ExonUTR.txt

!head 2019-05-29-All5xCpGs-ExonUTR.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {exonUTR} \
| wc -l
!echo "methylated loci overlaps with exon UTR"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {exonUTR} \
> 2019-05-29-MethLoci-ExonUTR.txt

!head 2019-05-29-MethLoci-ExonUTR.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {exonUTR} \
| wc -l
!echo "sparsely methylated loci overlaps with exon UTR"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {exonUTR} \
> 2019-05-29-SparseMethLoci-ExonUTR.txt

!head 2019-05-29-SparseMethLoci-ExonUTR.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {exonUTR} \
| wc -l
!echo "unmethylated loci overlaps with exon UTR"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {exonUTR} \
> 2019-05-29-UnMethLoci-ExonUTR.txt

!head 2019-05-29-UnMethLoci-ExonUTR.txt

### 4f. Genes

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {geneList} \
| wc -l
!echo "all 5x CpG loci overlaps with genes"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {geneList} \
> 2019-05-29-All5xCpGs-Genes.txt

!head 2019-05-29-All5xCpGs-Genes.txt

!cut -f6 2019-05-29-All5xCpGs-Genes.txt| sort | uniq -c | wc -l
!echo "unique genes represented in overlaps"

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {geneList} \
| wc -l
!echo "methylated loci overlaps with genes"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {geneList} \
> 2019-05-29-MethLoci-Genes.txt

!head 2019-05-29-MethLoci-Genes.txt

!cut -f6 2019-05-29-MethLoci-Genes.txt| sort | uniq -c | wc -l
!echo "unique genes represented in overlaps"

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {geneList} \
| wc -l
!echo "sparsely methylated loci overlaps with genes"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {geneList} \
> 2019-05-29-SparseMethLoci-Genes.txt

!head 2019-05-29-SparseMethLoci-Genes.txt

!cut -f6 2019-05-29-SparseMethLoci-Genes.txt| sort | uniq -c | wc -l
!echo "unique genes represesnted in overlaps"

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {geneList} \
| wc -l
!echo "unmethylated loci overlaps with genes"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {geneList} \
> 2019-05-29-UnMethLoci-Genes.txt

!head 2019-05-29-UnMethLoci-Genes.txt

!cut -f6 2019-05-29-UnMethLoci-Genes.txt| sort | uniq -c | wc -l
!echo "unique genes represented in overlaps"

### 4g. Putative promoters

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {putativePromoters} \
| wc -l
!echo "all 5x CpG loci overlaps with putative promoters"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {putativePromoters} \
> 2019-05-29-All5xCpGs-Putative-Promoters.txt

!head 2019-05-29-All5xCpGs-Putative-Promoters.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {putativePromoters} \
| wc -l
!echo "methylated loci overlaps with putative promoters"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {putativePromoters} \
> 2019-05-29-MethLoci-Putative-Promoters.txt

!head 2019-05-29-MethLoci-Putative-Promoters.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {putativePromoters} \
| wc -l
!echo "sparsely methylated loci overlaps with putative promoters"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {putativePromoters} \
> 2019-05-29-SparseMethLoci-Putative-Promoters.txt

!head 2019-05-29-SparseMethLoci-Putative-Promoters.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {putativePromoters} \
| wc -l
!echo "unmethylated loci overlaps with putative promoters"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {putativePromoters} \
> 2019-05-29-UnMethLoci-Putative-Promoters.txt

!head 2019-05-29-UnMethLoci-Putative-Promoters.txt

### 4h. Transposable elements (all)

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {transposableElementsAll} \
| wc -l
!echo "all 5x CpG loci overlaps with transposable elements (all)"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {transposableElementsAll} \
> 2019-05-29-All5xCpGs-TE-All.txt

!head 2019-05-29-All5xCpGs-TE-All.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {transposableElementsAll} \
| wc -l
!echo "methylated loci overlaps with transposable elements (all)"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {transposableElementsAll} \
> 2019-05-29-MethLoci-TE-All.txt

!head 2019-05-29-MethLoci-TE-All.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {transposableElementsAll} \
| wc -l
!echo "sparsely methylated loci overlaps with transposable elements (all)"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {transposableElementsAll} \
> 2019-05-29-SparseMethLoci-TE-All.txt

!head 2019-05-29-SparseMethLoci-TE-All.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {transposableElementsAll} \
| wc -l
!echo "unmethylated loci overlaps with transposable elements (all)"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {transposableElementsAll} \
> 2019-05-29-UnMethLoci-TE-All.txt

!head 2019-05-29-UnMethLoci-TE-All.txt

### 4i. Intergenic regions

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-u \
-a {all5xCpGs} \
-b {intergenic} \
| wc -l
!echo "all 5x CpG loci overlaps with intergenic regions"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {all5xCpGs} \
-b {intergenic} \
> 2019-05-29-All5xCpGs-intergenic.txt

!head 2019-05-29-All5xCpGs-intergenic.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylatedLoci} \
-b {intergenic} \
| wc -l
!echo "methylated loci overlaps with intergenic regions"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {methylatedLoci} \
-b {intergenic} \
> 2019-05-29-MethLoci-intergenic.txt

!head 2019-05-29-MethLoci-intergenic.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {sparselyMethylatedLoci} \
-b {intergenic} \
| wc -l
!echo "sparsely methylated loci overlaps with intergenic regions"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {sparselyMethylatedLoci} \
-b {intergenic} \
> 2019-05-29-SparseMethLoci-intergenic.txt

!head 2019-05-29-SparseMethLoci-intergenic.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-u \
-a {unmethylatedLoci} \
-b {intergenic} \
| wc -l
!echo "unmethylated loci overlaps with intergenic regions"

! {bedtoolsDirectory}intersectBed \
-wb \
-a {unmethylatedLoci} \
-b {intergenic} \
> 2019-05-29-UnMethLoci-intergenic.txt

!head 2019-05-29-UnMethLoci-intergenic.txt

### 4j. No overlaps

#### All 5x CpGs

! {bedtoolsDirectory}intersectBed \
-v \
-a {all5xCpGs} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
| wc -l
!echo "all 5x CpG loci do not overlap with exons, introns, transposable elements (all), or putative promoters"

! {bedtoolsDirectory}intersectBed \
-v \
-a {all5xCpGs} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
> 2019-05-29-All5xCpGs-NoOverlaps.txt

!head 2019-05-29-All5xCpGs-NoOverlaps.txt

#### Methylated loci

! {bedtoolsDirectory}intersectBed \
-v \
-a {methylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
| wc -l
!echo "methylated loci do not overlap with exons, introns, transposable elements (all), or putative promoters"

! {bedtoolsDirectory}intersectBed \
-v \
-a {methylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
> 2019-05-29-MethLoci-NoOverlaps.txt

!head 2019-05-29-MethLoci-NoOverlaps.txt

#### Sparsely methylated loci

! {bedtoolsDirectory}intersectBed \
-v \
-a {sparselyMethylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
| wc -l
!echo "sparsely methylated loci do not overlap with exons, introns, transposable elements (all), or putative promoters"

! {bedtoolsDirectory}intersectBed \
-v \
-a {sparselyMethylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
> 2019-05-29-SparseMethLoci-NoOverlaps.txt

!head 2019-05-29-SparseMethLoci-NoOverlaps.txt

#### Unmethylated loci

! {bedtoolsDirectory}intersectBed \
-v \
-a {unmethylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
| wc -l
!echo "unmethylated loci do not overlap with exons, introns, transposable elements (all), or putative promoters"

! {bedtoolsDirectory}intersectBed \
-v \
-a {unmethylatedLoci} \
-b {exonList} {intronList} {transposableElementsAll} {putativePromoters} \
> 2019-05-29-UnMethLoci-NoOverlaps.txt

!head 2019-05-29-UnMethLoci-NoOverlaps.txt

## 5. Identify methylation islands

To identify methylation islands using the method from Jeong et al. (2018), I need to define:

- starting size of the methylation window: 500 bp
- minimum fraction of methylated CpGs required within the window to be accepted: 0.02
- step size to extend the accepted window as long as the mCpG fraction is met: 50 bp
- mCpG file: input with mCpG chromosome and bp position

### 5a. Create mCpG file

#Modify mCpG file by removing the third column that is not needed for methylation island analysis
!awk '{print $1"\t"$2}' 2019-04-09-All-5x-CpG-Loci-Methylated.bed > 2019-04-09-All-5x-CpG-Loci-Methylated-Reduced.bed

#Confirm file only has chromosome and start bp for mCpG
!head 2019-04-09-All-5x-CpG-Loci-Methylated-Reduced.bed

### 5b. Identify methylation islands

#Identify methylation islands using 0.02 mCpG fraction (same as original paper)
! ./methyl_island_sliding_window.pl 500 0.02 50 2019-04-09-All-5x-CpG-Loci-Methylated-Reduced.bed \
> 2020-02-06-Methylation-Islands-500_0.02_50.tab

#Count max mCpG in an island
#Count min mCpG in an island
!awk 'NR==1{max = $4 + 0; next} {if ($4 > max) max = $4;} END {print max}' \
2020-02-06-Methylation-Islands-500_0.02_50.tab
!awk 'NR==1{min = $4 + 0; next} {if ($4 < min) min = $4;} END {print min}' \
2020-02-06-Methylation-Islands-500_0.02_50.tab

#Filter by MI length and print MI length in a new column
!awk '{if ($3-$2 >= 500) { print $1"\t"$2"\t"$3"\t"$4"\t"$3-$2}}' 2020-02-06-Methylation-Islands-500_0.02_50.tab \
> 2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab
!head 2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab
! wc -l 2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab

#Count max mCpG in an island
#Count min mCpG in an island
!awk 'NR==1{max = $4 + 0; next} {if ($4 > max) max = $4;} END {print max}' \
2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab
!awk 'NR==1{min = $4 + 0; next} {if ($4 < min) min = $4;} END {print min}' \
2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab

### 5c. Create BEDfiles for IGV and `bedtools`

#Create tab-delimited BEDfile without additional information
!awk '{print $1"\t"$2"\t"$3}' 2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab \
> 2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab.bed

#Check output
!head 2020-02-06-Methylation-Islands-200_0.02_50.tab.bed

### 5d. Identify genome feature overlaps with methylation islands

methylationIslands = "2020-02-06-Methylation-Islands-500_0.02_50-filtered.tab.bed"

!head {methylationIslands}
!wc -l {methylationIslands}

#### Genes

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylationIslands} \
-b {geneList} \
| wc -l
!echo "MI overlaps with genes"

! {bedtoolsDirectory}intersectBed \
-wo \
-a {methylationIslands} \
-b {geneList} \
> 2020-02-06-MI-Genes.txt

!head 2020-02-06-MI-Genes.txt

#### Intergenic regions

! {bedtoolsDirectory}intersectBed \
-u \
-a {methylationIslands} \
-b {intergenic} \
| wc -l
!echo "MI overlaps with intergenic regions"

! {bedtoolsDirectory}intersectBed \
-wo \
-a {methylationIslands} \
-b {intergenic} \
> 2020-02-06-MI-intergenic.txt

!head 2020-02-06-MI-intergenic.txt
